refactored utils function to a separate file
