+++ src/Game/Game.js
@@ -5,17 +5,9 @@ import Keyboard from '../Keyboard/Keyboard';
 import * as Clipboard from 'expo-clipboard';
 import WordList from '../words';
 import { styles } from './styles';
+import { copyBidirectionalArr, getDayOfTheYear } from '../utils';
 
 const NUMBER_OF_TRIES = 6;
-
-const getDayOfTheYear = () => {
-  const now = new Date();
-  const start = new Date(now.getFullYear(), 0, 0);
-  const diff = now - start;
-  const oneDay = 1000 * 60 * 60 * 24;
-  const day = Math.floor(diff / oneDay);
-  return day;
-};
 
 const wordOfTheDay = WordList[getDayOfTheYear()];
 
@@ -42,8 +34,6 @@ const Game = () => {
       checkGameState();
     }
   }, [currRow]);
-
-  const copyBidirectionalArr = (arr) => [...arr.map((row) => [...row])];
 
   const onKeyPressed = (key) => {
     if (gameState !== 'playing') return;
